pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pysensorlinx"          
version = "0.1.2"             
description = "Python library for accessing SensorLinx Device Data"
readme = "README.md"          
license = { text = "MIT" }    
authors = [
  { name = "Stefan Slivinski", email = "sslivins@gmail.com" }
]
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
# Runtime dependencies
dependencies = [
    "aiohttp>=3.11.12"
]

[project.optional-dependencies]
tests = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.3",
    "aioresponses>=0.7.8",
    "python-dotenv>=1.0.1"
]

[project.urls]
"Homepage" = "https://github.com/sslivins/pysensorlinx"
"Bug Tracker" = "https://github.com/sslivins/pysensorlinx/issues"

[tool.setuptools]
# Instruct setuptools to look in the 'src' folder.
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
